third_party/blink/web_tests/external/wpt/css/css-inline/baseline-shift/baseline-shift-sub-super-expected.txt

This is a testharness.js-based test.
[FAIL] baseline-shift: sub is below baseline
  assert_greater_than: expected a number greater than 7 but got 7
[FAIL] baseline-shift: super is above baseline
  assert_less_than: expected a number less than 7 but got 7
Harness: the test ran to completion.
